Spring-Boot-REST-Data-JPA-And-Data-Rest
We created an employees project that covers rest control and services in depth in Spring Boot and connected it to MySQL Workbench. We performed CRUD operations on Postman. Later, we saw the ease of use with data JPA and data Rest, we took advantage of spring's features and saved hundreds of lines of code.